requirements.txt
absl-py==1.0.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
autopep8==2.0.2
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
dnspython==2.2.1
email-validator==1.2.1
entrypoints==0.4
executing==0.8.3
fastapi==0.78.0
fastjsonschema==2.15.3
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
google-api-core==2.8.2
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.3.0
google-cloud-bigquery-storage==2.14.1
google-cloud-core==2.3.2
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.13.0
h5py==3.6.0
htmlmin==0.1.12
httptools==0.4.0
idna==3.3
ImageHash==4.3.1
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.7
jsonschema==4.5.1
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.4.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.13.0
jupyterlab-widgets==3.0.7
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
lxml==4.8.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
multimethod==1.9.1
nbclassic==0.3.7
nbclient==0.6.2
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.0
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
orjson==3.7.1
packaging==21.3
pandas==1.4.2
pandas-datareader==0.10.0
pandas-profiling==3.6.6
pandoc==2.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
phik==0.12.3
pickleshare==0.7.5
Pillow==9.1.0
plumbum==1.7.2
ply==3.11
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
proto-plus==1.20.6
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.8
PyPDF2==2.9.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyWavelets==1.4.1
pywin32==304
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scikit-plot==0.3.7
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
starlette==0.19.1
statsmodels==0.13.2
tangled-up-in-unicode==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-datasets==4.5.2
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-metadata==1.7.0
termcolor==1.1.0
terminado==0.13.3
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typeguard==2.13.3
typing_extensions==4.2.0
ujson==5.3.0
urllib3==1.26.9
uvicorn==0.17.6
visions==0.7.5
watchgod==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
websockets==10.3
Werkzeug==2.1.2
widgetsnbextension==4.0.7
wrapt==1.14.1
xgboost==1.6.1
ydata-profiling==4.1.2
zipp==3.8.0